Support for transparent statusbar.
This patch enables transparent statusbar in themes that support it,
and allows a different style for statusbar in 'date mode'.

There won't any visible changes to users who don't install these themes.

Patch 3: Now, The date is in the center of the status bar for better appearance.
Patch 4: Add mdpi devices

Change-Id: I22867350ce76c3e2793ca73850dcf3f35485ffad

Conflicts:

	packages/SystemUI/res/layout/status_bar.xml