Override external display dimensions for primary TV
OMAP_ENHANCEMENT

If we're using an HDMI primary display the PhoneWindowManager
still is aware that there is an external display and tries to
get the external width and height.

TI already has a patch to calculate a mirroring region if the
customer wants to specify a sub-region of the display to
present on the external display.

However, for a primary HDMI display it seems we shouldn't need
this and should present just the actual display size.

The platform needs to set the Android property "persist.hwc.primary.tv"
to "true" for this to take effect.

Change-Id: I5a6ec1ab4902c72537bb3feb4788284993bd37a4
Signed-off-by: Tony Lofthouse <a0741364@ti.com>