TabletStatusBar: Render GPU to improve render perf
OMAP_ENHANCEMENT

The StatusBar is normally drawn with the CPU, as mentioned it picks this
format because this is supported by the hardware composer (actually the DSS)
because if drew with the CPU Skia would be used and the format would be either
RGBA or RGBX which isn't supported by the OMAP DSS.

Problem with this, is this only works with 0 degree rotation. When CPU rendering
any rotated (UI) layer would have to be composited with SurfaceFlinger. This
composition is expensive, if we hardware accelerate the StatusBar this means
drawing the layer with the GPU and the trick here is that the EGL driver will
perform any rotation *before* it hits SurfaceFlinger. This means SurfaceFlinger
does not have to draw anything.

The expensive part here is that having SurfaceFlinger use the GPU to draw
as well applications using the GPU is more costly that getting the SystemUI to
exploit GPU rotation itself.

Change-Id: I3033dfe4b52a81f9ead4b0268a6bfb2d1fd4d3dc
Signed-off-by: Tony Lofthouse <a0741364@ti.com>