Limit maximum wallpaper size
OMAP_ENHANCEMENT

This patch addresses the issue of wallpapers using textures too
large for SGX to handle.

In the case of SGX540, a texture cannot be larger than 2048x2048
pixels. The Launcher2 app needs some extra width for wallpaper
parallax effect, and when the framebuffer is set to 1080p the
requested wallpaper size goes beyond the SGX limit. As a result,
the wallpaper manager service fails and SystemUI app crashes.

Change-Id: I2e05bf00eb1e355160d617900f5589880bfb7e10
Signed-off-by: Tony Lofthouse <a0741364@ti.com>